WIP: Added VCR cassettes around features
