+++ features/support/env.rb
@@ -57,9 +57,3 @@ end
 # The :transaction strategy is faster, but might give you threading problems.
 # See https://github.com/cucumber/cucumber-rails/blob/master/features/choose_javascript_database_strategy.feature
 Cucumber::Rails::Database.javascript_strategy = :truncation
-
-Capybara.register_driver :selenium do |app|
-  profile = Selenium::WebDriver::Firefox::Profile.new
-  profile.native_events = true
-  Capybara::Selenium::Driver.new(app, :browser => :firefox, profile: profile)
-end
